+++ examples/general/delete_game_server/main.go
