+++ examples/general/get_game_server/main.go
