+++ examples/general/get_server_metrics/main.go
